Reorganize CLI code into library
