+++ test/test_db.py
@@ -1,8 +1,8 @@
 from os import listdir, mkdir
 
-from cli.add import add
 from imgdb.config import Config, g_config
 from imgdb.db import *
+from imgdb.main import add
 
 IMGS = listdir('test/pics')
